+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.3.1] 2026-08-17
+
+## Changed
+
+* Added an interface for modifying the debounce time in both the Watcher::File object and monitoring Watcher::Panopticon
+
+## Bugfix
+
+* Overhauled the unit testing modules for the Async module so that they are less fragile. Your testing suite shouldn't have bugs!
+
 ## [3.3.0] 2026-08-16
 
 ## New Module : Async